Add Nature's Claim

+++ lib/magic/effects/destroy_controller_gains_life.rb
@@ -1,6 +1,7 @@
 module Magic
   module Effects
     class DestroyControllerGainsLife
+      attr_reader :valid_targets
       def initialize(valid_targets:)
         @valid_targets = valid_targets
       end
@@ -14,8 +15,10 @@ module Magic
       end
 
       def resolve(target:)
-        target.destroy!
-        target.controller.gain_life(4)
+        if valid_targets.call(target)
+          target.destroy!
+          target.controller.gain_life(4)
+        end
       end
     end
   end
